Update servertest.ini
Stoves blowing up randomly is really bad if you want to have a permanent home. Our first night on the server our house burned down for no reason other than frustrating random chance. This setting can be turned on and off on the fly by an admin in the panel too.